AI-Automation-Orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-automation-orchestrator"
version = "1.0.0"
description = "Unified platform combining OpenHands, crewAI, AutoGPT, Strix, and MoneyPrinterV2"
authors = [{name = "GENTS HUB", email = "gentshub@example.com"}]
readme = "README.md"
requires-python = ">=3.10,<3.15"
license = {text = "MIT"}
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "asyncio>=3.4.3",
    "typing-extensions>=4.0.0",
    "colorama>=0.4.6",
    "tabulate>=0.9.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0.0"
]
web = ["streamlit>=1.28.0"]
all = ["ai-automation-orchestrator[dev,web]"]

[project.scripts]
aiauto = "orchestrator.cli.main:cli"
aiauto-web = "orchestrator.api.server:main"

[tool.setuptools.packages.find]
include = ["orchestrator*"]
exclude = ["tests*", "examples*"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
